# src/main/java/com/junjie/rag/controller/UserController.java
package com.junjie.rag.controller;

import com.junjie.rag.common.ApplicationConstant;
import com.junjie.rag.common.BaseResponse;
import com.junjie.rag.common.PageResult;
import com.junjie.rag.common.ResultUtils;
import com.junjie.rag.constant.JwtClaimsConstant;
import com.junjie.rag.entity.User;
import com.junjie.rag.pojo.dto.PasswordDTO;
import com.junjie.rag.pojo.dto.UserDTO;
import com.junjie.rag.pojo.dto.UserPageQueryDTO;
import com.junjie.rag.pojo.vo.UserLoginVO;
import com.junjie.rag.config.JwtProperties;
import com.junjie.rag.service.UserService;
import com.junjie.rag.utils.JwtUtil;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.tags.Tag;
import lombok.extern.slf4j.Slf4j;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.util.DigestUtils;
import org.springframework.web.bind.annotation.*;

import javax.security.auth.login.AccountLockedException;
import javax.security.auth.login.AccountNotFoundException;
import java.util.HashMap;
import java.util.Map;

/**
 * @Title: UserController
 * @Author junjie
 * @Package com.junjie.rag.controller
 * @description: 用户控制层
 */

@Tag(name="UserController",description = "用户管理")
@Slf4j
@RestController
@RequestMapping(ApplicationConstant.API_VERSION + "/user")
public class UserController {
    @Autowired
    private UserService userService;
    @Autowired
    private JwtProperties jwtProperties;


    /**
     * 修改密码
     */
    @PostMapping("/updatePassword")
    @Operation(summary = "updatePassword",description = "修改密码")
    public BaseResponse updatePassword(@RequestBody PasswordDTO passwordDTO) {
        log.info("修改密码：{}", passwordDTO.toString());
        if (!passwordDTO.getNewPassword().equals(passwordDTO.getConfirmPassword())) {
            return ResultUtils.error("新密码与确认密码不一致");
        }
        User user = userService.getById(passwordDTO.getId());
        String s = DigestUtils.md5DigestAsHex(passwordDTO.getOldPassword().getBytes());
        if (!user.getPassword().equals(s)) {
            return ResultUtils.error("旧密码错误");
        }
        user.setPassword(DigestUtils.md5DigestAsHex(passwordDTO.getNewPassword().getBytes()));
        userService.updateById(user);
        return ResultUtils.success("修改密码成功");
    }

    /**
     * 注册
     *
     */
    @PostMapping("/register")
    @Operation(summary = "register",description = "注册")
    public BaseResponse register(@RequestBody User user) {
        log.info("注册：{}", user.toString());

        if (userService.getByUsername(user.getUserName())) {
            return ResultUtils.error("用户名已存在");
        }else {
            userService.register(user);
        }
        return ResultUtils.success("注册成功");
    }


    /**
     * 登录
     *
     * @param
     * @return
     */
    @PostMapping("/login")
    @Operation(summary = "login",description = "登录")
    public BaseResponse login(@RequestParam(value = "userName", defaultValue = "admin") String userName,
                                               @RequestParam(value = "password", defaultValue = "123456") String password) throws AccountLockedException, AccountNotFoundException {
        log.info("登录：{}", userName+":"+password);

        User user = userService.login(userName, password);

        //登录成功后，生成jwt令牌
        Map<String, Object> claims = new HashMap<>();
        claims.put(JwtClaimsConstant.USER_ID, user.getId());
        String token = JwtUtil.createJWT(
                jwtProperties.getUserSecretKey(),
                jwtProperties.getUserTtl(),
                claims);

        UserLoginVO userLoginVO = UserLoginVO.builder()
                .id(user.getId())
                .userName(user.getUserName())
                .name(user.getName())
                .token(token)
                .build();

        return ResultUtils.success(userLoginVO);
    }

    /**
     * 退出
     *
     * @return
     */
    @PostMapping("/logout")
    @Operation(summary = "logout",description = "退出")
    public BaseResponse<String> logout() {
        return ResultUtils.success("退出成功");
    }

    /**
     * 新增
     * @param userDTO
     * @return
     */
    @PostMapping("/addUser")
    @Operation(summary = "logout",description = "新增user")
    public BaseResponse save(@RequestBody UserDTO userDTO){
        log.info("新增员工：{}",userDTO);
        userService.saveUser(userDTO);
        return ResultUtils.success("新增成功");
    }

    /**
     * 员工分页查询
     * @param userPageQueryDTO
     * @return
     */
    @GetMapping("/page")
    @Operation(summary = "page",description = "user分页查询")
    public BaseResponse<PageResult> page(UserPageQueryDTO userPageQueryDTO){
        log.info("员工分页查询，参数为：{}", userPageQueryDTO);
        PageResult pageResult = userService.pageQuery(userPageQueryDTO);
        return ResultUtils.success(pageResult);
    }

    /**
     * 启用禁用员工账号
     * @param status
     * @param id
     * @return
     */
    @PostMapping("/status/{status}")
    @Operation(summary = "status",description = "启用禁用账号")
    public BaseResponse startOrStop(@PathVariable Integer status,Integer id){
        log.info("启用禁用员工账号：{},{}",status,id);
        userService.startOrStop(status,id);
        return ResultUtils.success("禁用成功");
    }

    /**
     * 根据id查询员工信息
     * @param id
     * @return
     */
    @GetMapping("/{id}")
    @Operation(summary = "info",description = "根据id查询user信息")
    public BaseResponse<User> getById(@PathVariable Long id){
        User employee = userService.getById(id);
        return ResultUtils.success(employee);
    }

    /**
     * 编辑员工信息
     * @param user
     * @return
     */
    @PutMapping("/update")
    @Operation(summary = "info",description = "编辑user信息")
    public BaseResponse update(@RequestBody User user){
        log.info("编辑员工信息：{}", user);
        userService.updateById(user);
        return ResultUtils.success("编辑成功");
    }
}